Added config.plist
This config.plist is configure based on this guide:
https://dortania.github.io/OpenCore-Desktop-Guide/config.plist/skylake.html

Also, there is 2 tricky parts that I had to do:
1. We need to swap these two values ShrinkMemoryMap/RebuildAppleMemoryMap
and SetupVirtualMap on the config.plist otherwise it won't boot into the
installer
2. Cube i7 Book uses iGPU so we need to configure with iGPU in the
config.plist, I followed this guide.
https://1revenger1.gitbook.io/laptop-guide/prepare-install-macos/display-configuration#intel-skylake

After all that, everytime you made a change to the kext or driver you need
to use ProperTree to Snapshot OC to re-populate the driver and kext back
into config.plist.

Signed-off-by: Stanley Phoong <stanley.phoong@gmail.com>