Fixed Docking keyboard and USB, remove VoodooI2C
Despite the guide stated that USBInjectAll kext is only for legacy
machines, but somehow Cube i7 Book Dock and keyboard needs this.

VoodooI2C kext was also removed because it doesn't seem to enable our
trackpad or touchscreen. I might re-introduce this back if some other kext
needs this again.

Signed-off-by: Stanley Phoong <stanley.phoong@gmail.com>